+++ packages/yourloops/components/patient-data.tsx
@@ -25,9 +25,9 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import React, { useEffect } from 'react'
+import React, { useEffect, useMemo } from 'react'
 import bows from 'bows'
-import { useParams } from 'react-router-dom'
+import { useLocation, useNavigate, useParams } from 'react-router-dom'
 import { useTranslation } from 'react-i18next'
 
 import { makeStyles } from 'tss-react/mui'
@@ -81,6 +81,10 @@ function PatientDataPage(): JSX.Element | null {
   const { getPatientById } = usePatientContext()
   const dataHook = useData()
   const { classes } = patientDataStyles()
+
+  const location = useLocation()
+  const navigate = useNavigate()
+  const pathName = useMemo(() => location.pathname, [location])
 
   const [patient, setPatient] = React.useState<Readonly<Patient> | null>(null)
   const [error, setError] = React.useState<string | null>(null)
@@ -151,8 +155,10 @@ function PatientDataPage(): JSX.Element | null {
   return (
     <Container className={classes.container} maxWidth={false}>
       <Blip
-        config={appConfig}
         api={blipApi}
+        config={appConfig}
+        navigate={navigate}
+        pathName={pathName}
         patient={patient}
         setPatient={setPatient}
         prefixURL={prefixURL}